Split canonical storage mappers
Move mapper functions into focused entity, history, and reference modules
so each file stays under the project line-count limit.

Keep the original mappers module as a compatibility re-export for tests and callers.
